+++ src/components/Input/Input.tsx
@@ -1,0 +1,10 @@
+import { StyledInput } from "../LoginForm/style";
+
+interface iInputProps {
+  register: {};
+  placeholder: string;
+}
+
+export function Input({ register, placeholder }: iInputProps) {
+  return <StyledInput {...register} placeholder={placeholder} />;
+}
